+++ examples/python/sysctl.py
@@ -18,7 +18,8 @@ import os
 import sys
 
 sys.path.append(os.path.dirname(__file__) + "/../contrib/bindings/python")
-import pathrs
+from pathrs import procfs
+from pathrs.procfs import ProcfsHandle
 
 
 def bail(*args):
@@ -38,18 +39,18 @@ def sysctl_subpath(name: str) -> str:
     return "sys/" + name.replace(".", "/")
 
 
-PROCFS = pathrs.ProcfsHandle.new(unmasked=True)
+PROCFS = ProcfsHandle.new(unmasked=True)
 
 
 def sysctl_write(name: str, value: str) -> None:
     subpath = sysctl_subpath(name)
-    with PROCFS.open(pathrs.PROC_ROOT, subpath, "w") as f:
+    with PROCFS.open(procfs.PROC_ROOT, subpath, "w") as f:
         f.write(value)
 
 
 def sysctl_read(name: str, *, value_only: bool = False) -> None:
     subpath = sysctl_subpath(name)
-    with PROCFS.open(pathrs.PROC_ROOT, subpath, "r") as f:
+    with PROCFS.open(procfs.PROC_ROOT, subpath, "r") as f:
         value = chomp(f.read())
         if value_only:
             print(f"{value}")